Fix miss calling of the polymophic methods -- connect() -- of WiFiEspClient class.

+++ src/WiFiEspClient.cpp
@@ -60,14 +60,14 @@ size_t WiFiEspClient::println(const __FlashStringHelper *ifsh)
 
 int WiFiEspClient::connectSSL(const char* host, uint16_t port)
 {
-	return connect(host, port, SSL_MODE);
+	return connect(host, port, (uint8_t)SSL_MODE);
 }
 
 int WiFiEspClient::connectSSL(IPAddress ip, uint16_t port)
 {
 	char s[16];
 	sprintf_P(s, PSTR("%d.%d.%d.%d"), ip[0], ip[1], ip[2], ip[3]);
-	return connect(s, port, SSL_MODE);
+	return connect(s, port, (uint8_t)SSL_MODE);
 }
 
 int WiFiEspClient::connect(const char *host, uint16_t port, int timeout)
@@ -86,7 +86,7 @@ int WiFiEspClient::connect(IPAddress ip, uint16_t port, int timeout)
 
 int WiFiEspClient::connect(const char* host, uint16_t port)
 {
-    return connect(host, port, TCP_MODE);
+    return connect(host, port, (uint8_t)TCP_MODE);
 }
 
 int WiFiEspClient::connect(IPAddress ip, uint16_t port)
@@ -94,7 +94,7 @@ int WiFiEspClient::connect(IPAddress ip, uint16_t port)
 	char s[16];
 	sprintf_P(s, PSTR("%d.%d.%d.%d"), ip[0], ip[1], ip[2], ip[3]);
 
-	return connect(s, port, TCP_MODE);
+	return connect(s, port, (uint8_t)TCP_MODE);
 }
 
 /* Private method */
